Update project grantee & fix rate limit bug
