+++ packages/web-app/app/_server/idos/index.ts
@@ -1,60 +1,49 @@
-// adapted from https://github.dev/idos-network/idos-sdk-js/tree/main/apps/idos-example-dapp
-
 import * as Base64Codec from '@stablelib/base64';
-import { ethers } from 'ethers';
-import nacl from 'tweetnacl';
 import { ServerPublicInfo } from '../types';
-import { hexToBytes } from 'viem';
 import { getGrants, getProjectApplicants } from './grants';
 import { addressesListSchema, grantsSchema } from '../../_types/schemas';
 import { isValidGrant } from '@/app/_lib/utils';
 import { blockedCountries } from '../blocked-countries';
 import { idOSGrantee } from './idos-grantee';
+import { ENCRYPTION_KEY_PAIR, evmGrantee } from '../wallet';
 
 export interface idOSGrant {
   content: string;
   encryption_public_key: string;
 }
 
-const ENCRYPTION_SECRET_KEY = hexToBytes(
-  process.env.NEXT_CITIZEND_WALLET_PRIVATE_KEY,
-);
-
-const EVM_GRANTEE_PRIVATE_KEY = process.env.NEXT_CITIZEND_WALLET_PRIVATE_KEY;
-const ENCRYPTION_KEY_PAIR = nacl.box.keyPair.fromSecretKey(
-  ENCRYPTION_SECRET_KEY,
-);
-
-const EVM_NODE_URL = 'https://ethereum-sepolia.publicnode.com';
-const evmGrantee = new ethers.Wallet(
-  EVM_GRANTEE_PRIVATE_KEY,
-  new ethers.JsonRpcProvider(EVM_NODE_URL),
-);
-
 const userFilter = async (grantee: idOSGrantee, userAddress: string) => {
-  const grantsResult = await getGrants({
-    owner: userAddress,
-    grantee: evmGrantee.address,
-  });
-  const grants = grantsSchema
-    .parse(grantsResult)
-    ?.filter((grant) => isValidGrant(grant.lockedUntil));
+  try {
+    const grantsResult = await getGrants({
+      owner: userAddress,
+      grantee: evmGrantee.address,
+    });
+    const grants = grantsSchema
+      .parse(grantsResult)
+      ?.filter((grant) => isValidGrant(grant.lockedUntil));
 
-  if (!grants.length) return null;
+    if (!grants.length) return null;
 
-  const country = await grantee.fetchUserCountryFromSharedPlusCredential(
-    grants[0].dataId,
-  );
+    const country = await grantee.fetchUserCountryFromSharedPlusCredential(
+      grants[0].dataId,
+    );
 
-  if (!country) return null;
+    if (!country) return null;
 
-  const isBlockedCountry = blockedCountries.some(
-    (blockedCountry) => blockedCountry === country,
-  );
+    const isBlockedCountry = blockedCountries.some(
+      (blockedCountry) => blockedCountry === country,
+    );
 
-  if (isBlockedCountry) return null;
+    if (isBlockedCountry) return null;
 
-  return userAddress;
+    return userAddress;
+  } catch (error) {
+    if (error instanceof Error) {
+      console.error(error.message);
+    }
+  }
+
+  return null;
 };
 
 export const getAllowedProjectApplicants = async (projectAddress: string) => {
@@ -63,14 +52,18 @@ export const getAllowedProjectApplicants = async (projectAddress: string) => {
     const addresses = addressesListSchema.parse(applicantsResult);
     const grantee = await idOSGrantee.init({
       granteeSigner: evmGrantee,
-      encryptionSecret: ENCRYPTION_SECRET_KEY,
+      encryptionSecret: ENCRYPTION_KEY_PAIR.secretKey,
     });
+    const allowed = [];
 
-    const result = await Promise.all(
-      addresses.map(async (address) => userFilter(grantee, address)),
-    ).then((results) => results.filter((element) => element !== null));
+    for (const address of addresses) {
+      const result = await userFilter(grantee, address);
+      if (result !== null) {
+        allowed.push(result);
+      }
+    }
 
-    return result as string[];
+    return allowed as string[];
   } catch (error) {
     console.error(error);
